+++ apps/payments/.env.sample
@@ -15,3 +15,4 @@ TL_CLIENT_SECRET=client_secret
 TL_CERTIFICATE_ID=certificate_kid
 TL_PRIVATE_KEY=private_key_joined_lines
 BACKEND_URL=http://localhost:8001
+NEXT_PUBLIC_HOST_URL=http://localhost:3001/
