+++ apps/payments/Dockerfile
@@ -22,6 +22,7 @@ ARG NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY
 ENV NODE_ENV=production
 ENV LOG_LEVEL=trace
 ENV NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY=$NEXT_PUBLIC_STRIPE_PUBLISHABLE_KEY
+ENV NEXT_PUBLIC_HOST_URL=$NEXT_PUBLIC_HOST_URL
 
 RUN npm run build --workspace=apps/payments
