+++ pipeline-variables/dev.yml
@@ -3,3 +3,4 @@ variables:
   awsRegion: eu-west-1
   ecrEndpoint: 730335224023.dkr.ecr.eu-west-1.amazonaws.com
   nextPublicStripePublishableKey: pk_test_51OzFkUK1VL0xyyr2E3EvcEmzsAGLowKH3jlM4g3AhQwUjhlpxQvfjZizc5GRSAnNzf8HHwTdzhz2rRYS69nooxtg00ZF7kRvyN
+  nextPublicHostUrl: http://payments.dev.blocks.gov.ie
